Add IOU class with tests

Closes #16

See merge request it1901/groups-2020/gr2011/gr2011!6

+++ BarKeep/core/src/main/java/barkeep/Drink.java
@@ -65,7 +65,7 @@ public class Drink {
      * @param name
      * @param value
      */
-    public Drink(String name, float value)
+    public Drink(String name, double value)
     {
         this.name = name;
         this.value = value;
